requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pyserial==3.5
setuptools==80.9.0
tkinter-tooltip==3.1.2
wheel==0.45.1
